Rename workflow file
